binlog store format

+++ hustdb/db/server/network/hustdb_network_utils.h
@@ -9,8 +9,8 @@ namespace hustdb_network {
 
 typedef struct
 {
-    unsigned long start;
-    unsigned long end;
+    unsigned int start;
+    unsigned int end;
 } ip_t;
 
 typedef struct
